Edit functionality updated where user cannot edit price, location, date and delete if the current date is less than end date + show page updated

+++ controllers/campgrounds/deleteCampground.php
@@ -3,7 +3,7 @@ include('../../utils/connect.php');
 include('../../views/partials/flash.php');
 session_start();
 $camp_id = $_GET['id'];
-$query = "SELECT users.username FROM users JOIN campgrounds ON users.id = campgrounds.owner_id
+$query = "SELECT users.username,end_date FROM users JOIN campgrounds ON users.id = campgrounds.owner_id
 WHERE campgrounds.id = $camp_id";
 
 $row = mysqli_query($con, $query);
@@ -18,6 +18,16 @@ if (empty($_SESSION["username"])) {
     header('location:http://localhost:3000/views/errorView.php?message=Unauthroied access!!');
     exit();
 }
+
+$current_date = strtotime(date('Y-m-d'));
+$end_date = strtotime($result['end_date']);
+
+if ($current_date < $end_date) {
+    $_SESSION["flash"] = flash("Current date must be greater than end date to perform this action", true);
+    header("location:../../views/campgrounds/index.php");
+    exit();
+}
+
 $camp = "DELETE FROM `campgrounds` WHERE `id`=$camp_id";
 mysqli_query($con, $camp);
 $_SESSION["flash"] = flash('You have deleted the campground succesfully!', false);
